requirements.txt
fpdf2==2.8.7
kaleido==0.2.1.post1
numpy
openpyxl
pandas
plotly
streamlit
